# swagger.yaml
openapi: 3.0.0
info:
  title: Backend Training API
  description: Comprehensive API documentation for backend training project
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /bad-request:
    post:
      summary: Show 400 Bad Request Error
      description: When required data is missing
      tags:
        - Error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: "GG"
                email:
                  type: string
                  example: "GG@example.com"
      responses:
        "400":
          description: Bad Request - Missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created successfully"
  /Unauthorized:
    get:
      summary: Shows 401 Unauthorized Error
      description: When user is not logged in
      tags:
        - Error
      parameters:
        - in: header
          name: Authorization
          required: false
          schema:
            type: string
            example: "Bearer your-token-here"
      responses:
        "401":
          description: Unauthorized - Please login
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Please login first"
                  status:
                    type: number
                    example: 401
        "200":
          description: Success - User is logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Welcome! You are logged in"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: An error occurred
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
...